+++ web/bootstrap/bootstrapper.go
@@ -4,19 +4,21 @@ import (
 	"time"
 
 	"github.com/insamo/mvc/core"
-	"github.com/insamo/mvc/datasource"
 
 	"os"
 
-	"github.com/go-kivik/kivik"
+	"github.com/insamo/mvc/datasource/transactions/nosql"
+	"github.com/insamo/mvc/datasource/transactions/sql"
 	"github.com/kataras/iris"
 	"github.com/kataras/iris/context"
 	"github.com/rakanalh/scheduler"
 	"github.com/rakanalh/scheduler/storage"
 )
 
+// Configurator func
 type Configurator func(*Bootstrapper)
 
+// Bootstrapper struct
 type Bootstrapper struct {
 	*iris.Application
 	AppName            string
@@ -26,8 +28,8 @@ type Bootstrapper struct {
 	RequestLogFile     *os.File
 	DatabaseLogFile    *os.File
 	RequestContext     *context.Handler
-	TxFactory          map[string]datasource.TransactionFactory
-	CoachFactory       map[string]*kivik.Client
+	TxFactory          map[string]sql.TransactionFactory
+	NxFactory          map[string]nosql.TransactionFactory
 	Environment        core.Config
 	Scheduler          scheduler.Scheduler
 }
@@ -62,8 +64,8 @@ func (b *Bootstrapper) Bootstrap(environment core.Config) *Bootstrapper {
 	b.Environment = environment
 
 	// Initialize transaction map
-	b.TxFactory = make(map[string]datasource.TransactionFactory)
-	b.CoachFactory = make(map[string]*kivik.Client)
+	b.TxFactory = make(map[string]sql.TransactionFactory)
+	b.NxFactory = make(map[string]nosql.TransactionFactory)
 
 	// Initialize scheduler
 	s := storage.NewMemoryStorage()
@@ -73,6 +75,7 @@ func (b *Bootstrapper) Bootstrap(environment core.Config) *Bootstrapper {
 	return b
 }
 
+// Close bootstraper
 func (b *Bootstrapper) Close() {
 	for _, instance := range b.Environment.DatabaseInstances() {
 		b.TxFactory[instance].Close()
@@ -80,6 +83,8 @@ func (b *Bootstrapper) Close() {
 	b.RequestLogFile.Close()
 	b.DatabaseLogFile.Close()
 	b.ApplicationLogFile.Close()
+	b.Scheduler.Clear()
+	b.Scheduler.Stop()
 }
 
 // Listen starts the http server with the specified "addr".